Restructure access log
