+++ backend/internal/system/log/log_test.go
@@ -22,8 +22,6 @@ import (
 	"bytes"
 	"io"
 	"log/slog"
-	"net/http"
-	"net/http/httptest"
 	"os"
 	"sync"
 	"testing"
@@ -206,48 +204,6 @@ func (suite *LogTestSuite) TestLoggerWith() {
 	assert.Contains(suite.T(), output, "Context log message")
 }
 
-func (suite *LogTestSuite) TestAccessLogHandler() {
-	var buf bytes.Buffer
-
-	logger = nil
-	once = sync.Once{}
-
-	handlerOptions := &slog.HandlerOptions{
-		Level: slog.LevelDebug,
-	}
-	logHandler := slog.NewTextHandler(&buf, handlerOptions)
-	log := &Logger{
-		internal: slog.New(logHandler),
-	}
-
-	testHandler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusOK)
-		_, err := w.Write([]byte("OK"))
-		if err != nil {
-			suite.T().Errorf("Failed to write response: %v", err)
-		}
-	})
-
-	handler := AccessLogHandler(log, testHandler)
-
-	req := httptest.NewRequest("GET", "/test", nil)
-	req.RemoteAddr = "192.168.1.1:12345"
-
-	rr := httptest.NewRecorder()
-
-	// Call the handler
-	handler.ServeHTTP(rr, req)
-
-	// Verify response
-	assert.Equal(suite.T(), http.StatusOK, rr.Code)
-	assert.Equal(suite.T(), "OK", rr.Body.String())
-
-	output := buf.String()
-	assert.Contains(suite.T(), output, "192.168.1.1")
-	assert.Contains(suite.T(), output, "GET /test")
-	assert.Contains(suite.T(), output, "200")
-}
-
 func (suite *LogTestSuite) TestMaskString() {
 	testCases := []struct {
 		name     string
@@ -302,32 +258,4 @@ func (suite *LogTestSuite) TestConvertFields() {
 	assert.Contains(suite.T(), output, "string=value")
 	assert.Contains(suite.T(), output, "int=42")
 	assert.Contains(suite.T(), output, "bool=true")
-}
-
-func (suite *LogTestSuite) TestLoggingResponseWriter() {
-	rec := httptest.NewRecorder()
-	lrw := &loggingResponseWriter{
-		ResponseWriter: rec,
-		statusCode:     http.StatusOK,
-		size:           0,
-	}
-
-	// Test writing headers
-	lrw.WriteHeader(http.StatusNotFound)
-	assert.Equal(suite.T(), http.StatusNotFound, lrw.statusCode)
-
-	// Test writing content
-	n, err := lrw.Write([]byte("test content"))
-	assert.NoError(suite.T(), err)
-	assert.Equal(suite.T(), 12, n)
-	assert.Equal(suite.T(), 12, lrw.size)
-
-	// Write more content
-	n, err = lrw.Write([]byte(" more"))
-	assert.NoError(suite.T(), err)
-	assert.Equal(suite.T(), 5, n)
-	assert.Equal(suite.T(), 17, lrw.size) // 12 + 5
-
-	// Verify the actual content was written to the underlying ResponseWriter
-	assert.Equal(suite.T(), "test content more", rec.Body.String())
 }
